Unmount the Zygote process
Previous implementation of DenyList still suffers from the problem of inconsistent mount IDs, because some mount IDs might be used by other processes while we are calling the `setns` or `unshare` functions.
To fully solve this problem, we used the first application process to cache a clean and a root mount namespace in the daemon.

In order to clean the mount namespace of Zygote, we should do it in `nativeForkAndSpecialize_pre` before the `fork` call.

In the zygiskd daemon, we made the following changes to improve stability
1. open mount namespace file descriptors in the main thread of zygiskd to avoid the case where an ad-hoc new thread might be recycled later;
2. use a while loop to ensure that the pipe between the child process and the parent process work as expected during the `fork` call of `save_mount_namespace`, because the child process might read the value it has written into the pipe, which will hang the parent process forever

After unmounting the Zygote process, we need to re-mount modules for application processes that are not on the denylist.
Note that out of unknown reason, it is unstable to first set a root mount namespace for them and then call unmounts.
We are left to implement the `magical_mount` function.