Re-implement DenyList as namespace profiles
With the help of `setns` system call, we cache three different mount namespace profiles: Clean, Root and Module.
All of them are created in the daemon using the mount namespace of the first application process.

Moreover, by defining DenyList modes as namespace profiles, we can extend them easily in the future.