Fix racing issue when caching mount profiles (#12)
There are two calls of `update_mount_namespace` with parameter `zygiskd::MountNamespace::Clean`.
In some rare case, this causes a racing issue, for example, between the daemon requests of [com.android.systemui] and [webview_zygote].

To fully solve the problem of thread racing on `save_mount_namespace`, we should cache mount profiles in the main thread of daemon.