Improve DenyList function for LineageOS (#13)
From user bug reports, we see that many app processes in LineageOS are created using `nativeForkAndSpecialize`.
Hence, we end up cleaning the Zygote process for multiple times.
Current commit fixes this problem. Additionally, it improves the logging of daemon.

Note that for KernelSU, if `unmount by default` is on, then we won't see log entries of update_mount_namespace.